requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
async-timeout==5.0.1
attrs==25.3.0
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
DeepImageSearch==2.5
dnspython==2.7.0
exceptiongroup==1.3.0
fastapi==0.115.13
filelock==3.18.0
fonttools==4.58.4
frozenlist==1.7.0
fsspec==2025.5.1
h11==0.16.0
huggingface-hub==0.33.0
idna==3.10
Jinja2==3.1.6
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.3
motor==3.7.1
mpmath==1.3.0
multidict==6.5.1
networkx==3.4.2
numpy==1.26.4
packaging==25.0
pandas==2.3.0
pillow==11.2.1
propcache==0.3.2
pydantic==2.11.7
pydantic_core==2.33.2
pymongo==4.13.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
requests==2.32.4
safetensors==0.5.3
six==1.17.0
sniffio==1.3.1
starlette==0.46.2
sympy==1.14.0
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.34.3
yarl==1.20.1


-f https://download.pytorch.org/whl/cpu/torch_stable.html
faiss-cpu==1.11.0
python-multipart==0.0.20
timm==1.0.15
torch==2.2.2+cpu
torchaudio==2.2.2+cpu
torchvision==0.17.2+cpu 


